Increase number of cxns in pool
For posterity, the issue with the docs examples seems to be that much of
the docs example data creation code is running under various
transactions. However not all operations in that code passes the txn to
sequelize. So what can happen is every cxn is "busy" in the middle of a
txn, then code tries to run a query outside a txn, but no cxn is
available to service it, so it waits... which blocks all of the txns
from proceeding.

Previous commits fix a number of cases of this, but not everything -- I
think something else is going on too -- but 20 cxns isn't unreasonable
even on my rpi I think? The min seems to be about 17 before things
explode.